+++ app-backend/db/src/main/scala/CampaignDao.scala
@@ -1,14 +1,17 @@
 package com.rasterfoundry.database
 
 import com.rasterfoundry.database.Implicits._
+import com.rasterfoundry.database.types._
 import com.rasterfoundry.datamodel._
 
+import cats.data.OptionT
 import cats.implicits._
 import doobie._
 import doobie.implicits._
 import doobie.implicits.javasql._
 import doobie.postgres.implicits._
 
+import java.sql.Timestamp
 import java.util.UUID
 
 object CampaignDao extends Dao[Campaign] with ObjectPermissions[Campaign] {
@@ -179,11 +182,133 @@ object CampaignDao extends Dao[Campaign] with ObjectPermissions[Campaign] {
   def isActiveCampaign(id: UUID): ConnectionIO[Boolean] =
     query.filter(id).filter(fr"is_active = ${true}").exists
 
+  def getChildren(campaignId: UUID): ConnectionIO[List[Campaign]] =
+    query.filter(fr"parent_campaign_id = $campaignId").list
+
   def getCloneOwners(id: UUID): ConnectionIO[List[UserThin]] =
     for {
-      campaigns <- query.filter(fr"parent_campaign_id = $id").list
+      campaigns <- getChildren(id)
       userIds = campaigns.map(_.owner)
       users <- UserDao.getUsersByIds(userIds)
       userThins = users.map(u => UserThin.fromUser(u))
     } yield userThins
+
+  def randomReviewTask(
+      campaignId: UUID,
+      user: User
+  ): ConnectionIO[Option[Task.TaskFeatureWithCampaign]] = {
+    fr"""
+      WITH parent_tasks (id) AS (
+        SELECT t1.id
+        FROM tasks t1
+        WHERE t1.annotation_project_id IN (
+          SELECT id
+          FROM annotation_projects
+          WHERE campaign_id IN (
+            SELECT id
+            FROM campaigns c
+            WHERE c.parent_campaign_id = ${campaignId}
+          )
+        )
+        AND t1.task_type = ${TaskType.Label.toString}::task_type
+        AND t1.parent_task_id IS NULL
+      ),
+      -- ids of tasks that have at least 3 validated sub-tasks
+      validated_ids (id) AS (
+        SELECT t1.id
+        FROM parent_tasks t1
+        JOIN tasks t2
+          ON t2.parent_task_id = t1.id
+        WHERE t2.task_type = ${TaskType.Review.toString}::task_type
+          AND t2.status <> ${TaskStatus.Labeled.toString}::task_status
+        GROUP BY t1.id
+        HAVING COUNT(t2.id) >= 3
+      ),
+      reviewed_by_user_counts (id, id2, reviewed) AS (
+        SELECT t1.id, t2.id, COUNT(ta.*)
+        FROM parent_tasks t1
+        JOIN tasks t2
+          ON t2.parent_task_id = t1.id
+        JOIN task_actions ta
+          ON ta.task_id = t2.id
+          AND ta.user_id = ${user.id}
+          AND ta.to_status = ${TaskStatus.Validated.toString}::task_status
+        WHERE t2.task_type = ${TaskType.Review.toString}::task_type
+        GROUP BY t1.id, t2.id
+      ),
+      reviewed_ids (id) AS (
+        SELECT parent_tasks.id
+        FROM parent_tasks
+        JOIN reviewed_by_user_counts r
+          ON r.id = parent_tasks.id
+        GROUP BY parent_tasks.id
+        HAVING SUM(r.reviewed) >= 1
+      )
+      SELECT tasks.*, annotation_projects.campaign_id
+      FROM tasks
+      JOIN annotation_projects
+        ON tasks.annotation_project_id = annotation_projects.id
+      WHERE
+        parent_task_id NOT IN (SELECT id FROM validated_ids)
+        AND parent_task_id NOT IN (SELECT id FROM reviewed_ids)
+        AND parent_task_id IS NOT NULL
+        AND task_type = ${TaskType.Review.toString}::task_type
+      ORDER BY RANDOM() LIMIT 1;
+    """
+      .query[Task.TaskWithCampaign]
+      .to[List] flatMap { tasks =>
+      tasks.toNel traverse { tasks =>
+        (for {
+          task <- OptionT(TaskDao.getTaskById(tasks.head.id))
+          actions <- OptionT.liftF(TaskDao.getTaskActions(task.id))
+        } yield {
+          tasks.head.toGeoJSONFeature(actions)
+        }).value
+      }
+    } map { _.flatten }
+  }
+
+  def getProjectMapping(
+      childCampaignId: UUID,
+      parentDateMap: Map[Timestamp, UUID]
+  ): ConnectionIO[List[(ChildAnnotationProjectId, ParentAnnotationProjectId)]] =
+    AnnotationProjectDao.listByCampaign(childCampaignId) map { childProjects =>
+      childProjects flatMap { project =>
+        project.capturedAt flatMap { parentDateMap.get(_) } map { parentId =>
+          (
+            ChildAnnotationProjectId(project.id),
+            ParentAnnotationProjectId(parentId)
+          )
+        }
+      }
+    }
+
+  def retrieveChildCampaignAnnotations(
+      campaignId: UUID
+  ): ConnectionIO[Unit] =
+    for {
+      childCampaigns <- getChildren(campaignId)
+      parentProjects <- AnnotationProjectDao.listByCampaign(campaignId)
+      // create a map from capture dates to parent projects. we're assuming for
+      // now that capture date is unique within a campaign
+      captureDates = Map((parentProjects flatMap { project =>
+        project.capturedAt map { dt =>
+          (dt, project.id)
+        }
+      }): _*)
+      // for each campaign, build a map of its annotation projects to the parent's
+      // annotation projects
+      childParentMappings <- childCampaigns traverse { campaign =>
+        getProjectMapping(campaign.id, captureDates)
+      }
+      // mash all those maps together -- there should be no collisions, since annotation
+      // projects can only be in a single campaign
+      childParentMapping = childParentMappings.combineAll
+      // now we have a map of child annotation projects to parent annotation projects,
+      // so we no longer care about campaigns
+      _ <- childParentMapping traverse {
+        case (childId, parentId) =>
+          AnnotationLabelDao.copyProjectAnnotations(childId, parentId)
+      }
+    } yield ()
 }