+++ app-backend/db/src/test/scala/com/azavea/rf/database/CampaignDaoSpec.scala
@@ -1,21 +1,26 @@
 package com.rasterfoundry.database
 
 import com.rasterfoundry.common.Generators.Implicits._
+import com.rasterfoundry.database.Implicits._
 import com.rasterfoundry.datamodel._
 
 import cats.implicits._
 import doobie.implicits._
+import doobie.postgres.implicits._
 import org.scalacheck.Prop.forAll
 import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.matchers.should.Matchers
 import org.scalatestplus.scalacheck.Checkers
+
+import scala.util.Random
 
 class CampaignDaoSpec
     extends AnyFunSuite
     with Matchers
     with Checkers
     with DBTestConfig
-    with PropTestHelpers {
+    with PropTestHelpers
+    with ConnectionIOLogger {
   test("insert a campaign") {
     check {
       forAll(
@@ -509,6 +514,409 @@ class CampaignDaoSpec
           true
         }
       )
+    }
+  }
+
+  test("no tasks are returned when all have been validated") {
+    check {
+      forAll(
+        (
+            userCreateBase: User.Create,
+            userCreates: List[User.Create],
+            userCreate: User.Create,
+            campaignCreate: Campaign.Create,
+            annotationProjectCreate: AnnotationProject.Create,
+            taskFeaturesCreate: Task.TaskFeatureCollectionCreate
+        ) => {
+          val copyIO = for {
+            parent <- UserDao.create(userCreate)
+            children <- userCreateBase :: userCreates traverse { u =>
+              UserDao.create(u)
+            }
+            insertedCampaign <- CampaignDao
+              .insertCampaign(
+                campaignCreate.copy(parentCampaignId = None),
+                parent
+              )
+            insertedProject <- AnnotationProjectDao
+              .insert(
+                annotationProjectCreate.copy(
+                  campaignId = Some(insertedCampaign.id),
+                  status = AnnotationProjectStatus.Waiting
+                ),
+                parent
+              )
+            _ <- AnnotationProjectDao.update(
+              insertedProject.toProject
+                .copy(status = AnnotationProjectStatus.Ready),
+              insertedProject.id
+            )
+            campaignCopies <- children traverse { child =>
+              CampaignDao.copyCampaign(insertedCampaign.id, child)
+            }
+            projectCopies <- (campaignCopies traverse { c =>
+              AnnotationProjectDao.listByCampaign(c.id)
+            }) map (_.flatten)
+            projectCopiesWithRelated <- projectCopies traverse { pc =>
+              AnnotationProjectDao.getWithRelatedById(pc.id)
+            } map (_.flatten)
+            tasks <- TaskDao.insertTasks(
+              fixupTaskFeaturesCollection(
+                taskFeaturesCreate,
+                projectCopiesWithRelated.head
+              ),
+              parent
+            )
+            _ <- TaskDao.updateTask(
+              tasks.features.head.id,
+              Task.TaskFeatureCreate(
+                tasks.features.head.properties.toCreate
+                  .copy(
+                    status = TaskStatus.Labeled,
+                    parentTaskId = None,
+                    taskType = Some(TaskType.Label)
+                  ),
+                tasks.features.head.geometry,
+                "Feature"
+              ),
+              parent
+            )
+            _ <- tasks.features.tail traverse { f =>
+              TaskDao.updateTask(
+                f.id,
+                Task.TaskFeatureCreate(
+                  f.properties.toCreate
+                    .copy(
+                      status = TaskStatus.Validated,
+                      parentTaskId = Some(tasks.features.head.id),
+                      taskType = Some(TaskType.Review)
+                    ),
+                  f.geometry,
+                  "Feature"
+                ),
+                parent
+              )
+            }
+            randomTask <- CampaignDao
+              .randomReviewTask(insertedCampaign.id, parent)
+          } yield randomTask
+
+          val randomTask = copyIO.transact(xa).unsafeRunSync
+
+          assert(
+            randomTask.isEmpty,
+            "When all tasks are validated no tasks should be returned"
+          )
+
+          true
+        }
+      )
+    }
+  }
+
+  test("a task is returned when none have been validated") {
+    check {
+      forAll(
+        (
+            userCreate1: User.Create,
+            userCreate2: User.Create,
+            userCreates: List[User.Create],
+            userCreate: User.Create,
+            campaignCreate: Campaign.Create,
+            annotationProjectCreate: AnnotationProject.Create,
+            taskFeaturesCreate: Task.TaskFeatureCollectionCreate
+        ) => {
+          val copyIO = for {
+            parent <- UserDao.create(userCreate)
+            children <- userCreate1 :: userCreate2 :: userCreates traverse {
+              u =>
+                UserDao.create(u)
+            }
+            insertedCampaign <- CampaignDao
+              .insertCampaign(
+                campaignCreate.copy(parentCampaignId = None),
+                parent
+              )
+            insertedProject <- AnnotationProjectDao
+              .insert(
+                annotationProjectCreate.copy(
+                  campaignId = Some(insertedCampaign.id),
+                  status = AnnotationProjectStatus.Waiting
+                ),
+                parent
+              )
+            _ <- AnnotationProjectDao.update(
+              insertedProject.toProject
+                .copy(status = AnnotationProjectStatus.Ready),
+              insertedProject.id
+            )
+            campaignCopies <- children traverse { child =>
+              CampaignDao.copyCampaign(insertedCampaign.id, child)
+            }
+            projectCopies <- (campaignCopies traverse { c =>
+              AnnotationProjectDao.listByCampaign(c.id)
+            }) map (_.flatten)
+            projectCopiesWithRelated <- projectCopies traverse { pc =>
+              AnnotationProjectDao.getWithRelatedById(pc.id)
+            } map (_.flatten)
+            tasks <- TaskDao.insertTasks(
+              fixupTaskFeaturesCollection(
+                taskFeaturesCreate,
+                projectCopiesWithRelated.head
+              ),
+              parent
+            )
+            _ <- TaskDao.updateTask(
+              tasks.features.head.id,
+              Task.TaskFeatureCreate(
+                tasks.features.head.properties.toCreate
+                  .copy(
+                    status = TaskStatus.Labeled,
+                    taskType = Some(TaskType.Label)
+                  ),
+                tasks.features.head.geometry,
+                "Feature"
+              ),
+              parent
+            )
+            _ <- tasks.features.tail traverse { f =>
+              TaskDao.updateTask(
+                f.id,
+                Task.TaskFeatureCreate(
+                  f.properties.toCreate
+                    .copy(
+                      status = TaskStatus.Labeled,
+                      parentTaskId = Some(tasks.features.head.id),
+                      taskType = Some(TaskType.Review)
+                    ),
+                  f.geometry,
+                  "Feature"
+                ),
+                parent
+              )
+            }
+            randomTask <- CampaignDao
+              .randomReviewTask(insertedCampaign.id, parent)
+          } yield (tasks, randomTask)
+          val (tasks, randomTask) = copyIO.transact(xa).unsafeRunSync
+
+          assert(
+            tasks.features.length > 1 == randomTask.isDefined,
+            "When no tasks are validated one task should be returned"
+          )
+
+          true
+        }
+      )
+    }
+  }
+
+  test("a task is not returned when parent has been validated by user") {
+    check {
+      forAll {
+        (
+            userCreates: List[User.Create],
+            userCreates1: (User.Create, User.Create, User.Create),
+            campaignCreate: Campaign.Create,
+            annotationProjectCreate: AnnotationProject.Create,
+            taskFeatureCreates: (Task.TaskFeatureCreate, Task.TaskFeatureCreate),
+            taskFeatureCollectionCreate: Task.TaskFeatureCollectionCreate
+        ) =>
+          {
+            val (userCreate1, userCreate2, userCreate3) = userCreates1
+            val (taskFeatureCreate1, taskFeatureCreate2) = taskFeatureCreates
+            val copyIO = for {
+              parent <- UserDao.create(userCreate1)
+              children <- userCreate2 :: userCreate3 :: userCreates traverse {
+                u =>
+                  UserDao.create(u)
+              }
+              insertedCampaign <- CampaignDao
+                .insertCampaign(
+                  campaignCreate.copy(parentCampaignId = None),
+                  parent
+                )
+              insertedProject <- AnnotationProjectDao
+                .insert(
+                  annotationProjectCreate.copy(
+                    campaignId = Some(insertedCampaign.id),
+                    status = AnnotationProjectStatus.Waiting
+                  ),
+                  parent
+                )
+              _ <- AnnotationProjectDao.update(
+                insertedProject.toProject
+                  .copy(status = AnnotationProjectStatus.Ready),
+                insertedProject.id
+              )
+              campaignCopies <- children traverse { child =>
+                CampaignDao.copyCampaign(insertedCampaign.id, child)
+              }
+              projectCopies <- (campaignCopies traverse { c =>
+                AnnotationProjectDao.listByCampaign(c.id)
+              }) map (_.flatten)
+              projectCopiesWithRelated <- projectCopies traverse { pc =>
+                AnnotationProjectDao.getWithRelatedById(pc.id)
+              } map (_.flatten)
+              tasks <- TaskDao.insertTasks(
+                fixupTaskFeaturesCollection(
+                  Task.TaskFeatureCollectionCreate(
+                    features = taskFeatureCreate1 :: taskFeatureCreate2.copy(
+                      properties = taskFeatureCreate2.properties.copy(
+                        status = TaskStatus.Labeled
+                      )
+                    ) :: taskFeatureCollectionCreate.features
+                  ),
+                  projectCopiesWithRelated.head
+                ),
+                parent
+              )
+              _ <- TaskDao.updateTask(
+                tasks.features.head.id,
+                Task.TaskFeatureCreate(
+                  tasks.features.head.properties.toCreate
+                    .copy(
+                      status = TaskStatus.Labeled,
+                      taskType = Some(TaskType.Label)
+                    ),
+                  tasks.features.head.geometry,
+                  "Feature"
+                ),
+                parent
+              )
+              _ <- TaskDao.updateTask(
+                tasks.features(1).id,
+                Task.TaskFeatureCreate(
+                  tasks
+                    .features(1)
+                    .properties
+                    .toCreate
+                    .copy(
+                      status = TaskStatus.Validated,
+                      parentTaskId = Some(tasks.features.head.id),
+                      taskType = Some(TaskType.Review)
+                    ),
+                  tasks.features(1).geometry,
+                  "Feature"
+                ),
+                parent
+              )
+              _ <- tasks.features.drop(2) traverse { f =>
+                TaskDao.updateTask(
+                  f.id,
+                  Task.TaskFeatureCreate(
+                    f.properties.toCreate
+                      .copy(
+                        status = TaskStatus.Labeled,
+                        parentTaskId = Some(tasks.features.head.id),
+                        taskType = Some(TaskType.Review)
+                      ),
+                    f.geometry,
+                    "Feature"
+                  ),
+                  parent
+                )
+              }
+              randomTask <- CampaignDao
+                .randomReviewTask(insertedCampaign.id, parent)
+            } yield randomTask
+
+            val randomTask = copyIO.transact(xa).unsafeRunSync
+
+            assert(
+              randomTask.isEmpty,
+              "When a user has already validated a task, no tasks are returned"
+            )
+            true
+          }
+      }
+    }
+  }
+
+  test("copy annotations back from child campaigns") {
+    check {
+      forAll {
+        (
+            users: (User.Create, User.Create),
+            sourceCampaignCreate: Campaign.Create,
+            sourceAnnotationProject: AnnotationProject.Create,
+            sourceTaskFeature: Task.TaskFeatureCreate,
+            childLabels: List[AnnotationLabelWithClasses.Create]
+        ) =>
+          {
+            val (parentUser, childUser) = users
+            val retrievalIO = for {
+              dbParentUser <- UserDao.create(parentUser)
+              dbParentCampaign <- CampaignDao.insertCampaign(
+                sourceCampaignCreate.copy(parentCampaignId = None),
+                dbParentUser
+              )
+              dbParentAnnotationProject <- AnnotationProjectDao.insert(
+                sourceAnnotationProject.copy(
+                  campaignId = Some(dbParentCampaign.id)
+                ),
+                dbParentUser
+              )
+              _ <- TaskDao.insertTasks(
+                Task.TaskFeatureCollectionCreate(
+                  _type = "FeatureCollection",
+                  features = List(
+                    sourceTaskFeature.copy(
+                      properties = sourceTaskFeature.properties.copy(
+                        annotationProjectId = dbParentAnnotationProject.id
+                      )
+                    )
+                  )
+                ),
+                dbParentUser
+              )
+              dbChildUser <- UserDao.create(childUser)
+              dbChildCampaign <- CampaignDao.copyCampaign(
+                dbParentCampaign.id,
+                dbChildUser
+              )
+              childProject <- AnnotationProjectDao.listByCampaign(
+                dbChildCampaign.id
+              ) map { _.head }
+              childProjectWithRelated <- AnnotationProjectDao
+                .getWithRelatedById(childProject.id) map { _.get }
+              childClasses = childProjectWithRelated.labelClassGroups flatMap {
+                _.labelClasses
+              } map { _.id }
+              childTask <- TaskDao.query
+                .filter(fr"annotation_project_id = ${childProject.id}")
+                .select
+              childInserted <- AnnotationLabelDao.insertAnnotations(
+                childProject.id,
+                childTask.id,
+                childLabels map { labelCreate =>
+                  labelCreate.copy(
+                    annotationLabelClasses =
+                      Random.shuffle(childClasses).take(1)
+                  )
+                },
+                dbChildUser
+              )
+              _ <- CampaignDao.retrieveChildCampaignAnnotations(
+                dbParentCampaign.id
+              )
+              labelCountOnParentProject <- AnnotationLabelDao.query
+                .filter(
+                  fr"annotation_project_id = ${dbParentAnnotationProject.id}"
+                )
+                .count
+            } yield (childInserted, labelCountOnParentProject)
+
+            val (childInsertedLabels, parentLabelCount) =
+              retrievalIO.transact(xa).unsafeRunSync
+
+            assert(
+              parentLabelCount == childInsertedLabels.length,
+              "Parent project has the same count of labels as the child project"
+            )
+            true
+          }
+      }
     }
   }
 }